requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
langchain==0.1.12
packaging
openai==1.12.0
selenium==4.18.1
webdriver-manager==4.0.1 